feat: 完成 Sprint 14 發布 - M09 MQ通知 + M07 Stripe整合 + M18 FAQ進階
## Sprint 14 完成項目

### M09 通知系統 Phase 2
- NotificationProducerService (Redis Stream MQ)
- NotificationConsumerService (訊息消費者)
- NotificationService (MQ 非同步發送整合)
- RedisStreamConfig (串流配置)

### M07 金流 Phase 2-B
- PaymentGateway (統一介面)
- StripePaymentGateway (Stripe 實現)
- LinePayPaymentGateway (LinePay 實現)
- MockPaymentGateway (Mock 實現)
- StripeWebhookController (Webhook 處理)

### M18 FAQ 進階功能
- FaqArticleController (置頂文章、高亮搜尋)
- FaqCategoryController (分類統計)
- FaqService (業務邏輯更新)

### 程式碼品質優化
- 修復 Exception 類別 serialVersionUID 警告
- 修復 AnalyticsService redundant cast 警告
- 新增 maven-compiler-plugin 配置 (-Xlint:all)

+++ backend/src/main/java/com/nextkey/ecommerce/core/analytics/AnalyticsService.java
@@ -84,7 +84,7 @@ public class AnalyticsService {
                 tenantId, today.atStartOfDay(), today.atTime(LocalTime.MAX));
         int yesterdayOrders = orderRepository.countByTenantIdAndCreatedAtBetween(
                 tenantId, yesterday.atStartOfDay(), yesterday.atTime(LocalTime.MAX));
-        int monthOrders = (int) orderRepository.countByTenantIdAndCreatedAtBetween(
+        int monthOrders = orderRepository.countByTenantIdAndCreatedAtBetween(
                 tenantId, monthStart.atStartOfDay(), today.atTime(LocalTime.MAX));
 
         // 待處理訂單
@@ -93,7 +93,7 @@ public class AnalyticsService {
         // 活躍 listing 統計
         int activeProducts = productRepository.countByListingTenantIdAndListingStatus(
                 tenantId, Listing.ListingStatus.ACTIVE);
-        int totalRooms = (int) roomRepository.countByListingTenantId(tenantId);
+        int totalRooms = roomRepository.countByListingTenantId(tenantId);
 
         return AnalyticsDto.DashboardStats.builder()
                 .todayRevenue(todayRevenue)
